fix: remove imports from dev config & version override
